employee-service/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
pymongo==4.6.0
python-dotenv==1.0.0
kafka-python==2.0.2
prometheus-client==0.19.0
flask-swagger-ui==4.11.1
pydantic==2.5.0
bcrypt==4.1.1
